+++ src/screens/Authentication/Signout/Signout.tsx
@@ -1,0 +1,26 @@
+import React, { ReactElement, useContext } from 'react';
+import { View } from 'react-native';
+import { AuthContext } from '../../../context/AuthContext';
+import StyledButton from '../../../components/StyledButton/StyledButton';
+import styles from './styles';
+
+function HomeScreen({ navigation }: any): ReactElement {
+  // const { user } = useAuthentication();
+
+  const { signOut } = useContext(AuthContext);
+
+  return (
+    <View style={styles.container}>
+      <StyledButton
+        text="Sign Out"
+        onPress={signOut}
+        buttonStyle={{ width: '45%', height: '5%' }}
+        textStyle = {{}}
+        activeOpacity = {{}}
+      />
+    </View>
+  );
+}
+
+
+export default HomeScreen;
