+++ src/screens/Authentication/Signout/styles.tsx
@@ -1,0 +1,13 @@
+import { StyleSheet, Dimensions, PixelRatio } from 'react-native';
+
+export default StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: '#fff',
+        alignItems: 'center',
+        justifyContent: 'center'
+      },
+      button: {
+        marginTop: 10
+      }
+    });
